+++ src/Settings.php
@@ -51,6 +51,12 @@ class Settings
         self::infoText(__('Use this setting if you want to be able to style the date picker in contact forms.', 'recras'));
     }
 
+    public static function addInputFixDatepicker(array $args): void
+    {
+        self::addInputCheckbox($args);
+        self::infoText(__('On some websites, the date picker in a book process has a tiny font. Enable this to fix this.', 'recras'));
+    }
+
     /**
      * Add a decimal separator input field
      */
@@ -347,6 +353,7 @@ class Settings
         self::registerSetting('recras_currency', '€');
         self::registerSetting('recras_decimal', ',');
         self::registerSetting('recras_datetimepicker', false, 'boolean');
+        self::registerSetting('recras_fix_react_datepicker', false, 'boolean');
         self::registerSetting('recras_theme', 'basic');
         self::registerSetting('recras_enable_analytics', false, 'boolean');
     }
@@ -366,6 +373,7 @@ class Settings
         self::addField('recras_currency', __('Currency symbol', 'recras'), [__CLASS__, 'addInputCurrency']);
         self::addField('recras_decimal', __('Decimal separator', 'recras'), [__CLASS__, 'addInputDecimal']);
         self::addField('recras_datetimepicker', __('Use calendar widget for contact forms', 'recras'), [__CLASS__, 'addInputDatepicker']);
+        self::addField('recras_fix_react_datepicker', __('Fix book process datepicker styling', 'recras'), [__CLASS__, 'addInputFixDatepicker']);
         self::addField('recras_theme', __('Theme for Recras integrations', 'recras'), [__CLASS__, 'addInputTheme']);
         self::addField('recras_enable_analytics', __('Enable Google Analytics integration? (deprecated)', 'recras'), [__CLASS__, 'addInputAnalytics']);
     }